norme

+++ ft_memmove.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_memmove.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: todina-r <todina-r@student.42antananarivo  +#+  +:+       +#+        */
+/*   By: todina-r <todina-r@student.42antananari    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/22 09:26:08 by todina-r          #+#    #+#             */
-/*   Updated: 2026/01/28 21:06:13 by todina-r         ###   ########.fr       */
+/*   Updated: 2026/02/04 11:43:18 by todina-r         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,12 +14,12 @@
 
 void	*ft_memmove(void *dest, const void *src, size_t n)
 {
-	int	index;
+	size_t	index;
 
 	if (dest < src)
 	{
 		index = 0;
-		while (index < (int)n)
+		while (index < n)
 		{
 			((char *)dest)[index] = ((char *)src)[index];
 			index++;
